Portability: Guard compiler builtins and fix integer width in jit module
- Wrap __builtin___clear_cache with __GNUC__/__clang__ guard; add
  FlushInstructionCache fallback for _WIN32 (portability-eval-bc-jit-001)
- Change stencil_desc_t .size/.nsymbols/.nrelocs from unsigned long to
  size_t; update related loop variables in emit.c (portability-eval-bc-jit-005)
- Change target_pc from long to ptrdiff_t in entry.c (portability-eval-bc-jit-003)
- Add MAP_ANONYMOUS fallback for MAP_ANON in region.c (portability-eval-bc-jit-006)

+++ src/eval/bc/jit/entry.c
@@ -728,7 +728,7 @@ static int prefix_has_escaping_branch(const mino_bc_fn_t *bc, size_t deopt_pc)
     for (size_t pc = 0; pc < deopt_pc; pc++) {
         unsigned op = OP_OF(bc->code[pc]);
         if (op != OP_JMP && op != OP_JMPIFNOT) continue;
-        long target_pc = (long)pc + 1 + sBx_OF(bc->code[pc]);
+        ptrdiff_t target_pc = (ptrdiff_t)pc + 1 + sBx_OF(bc->code[pc]);
         if (target_pc < 0 || (size_t)target_pc >= deopt_pc) return 1;
     }
     return 0;
